#54 fix name controller
