+++ src/main/java/cnpm/ergo/controller/Admin/Customer/DeleteCustomerController.java
@@ -1,7 +1,6 @@
-package cnpm.ergo.controller.Admin.Employee;
+package cnpm.ergo.controller.Admin.Customer;
 
-import cnpm.ergo.service.interfaces.IEmployeeService;
-import cnpm.ergo.service.implement.EmployeeServiceImpl;
+import cnpm.ergo.service.implement.CustomerServiceImpl;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -9,16 +8,18 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
+import cnpm.ergo.service.interfaces.ICustomerService;
 
-@WebServlet(name = "DeleteController", value = "/admin/employee/delete")
-public class DeleteController extends HttpServlet {
-    private IEmployeeService employeeService;
+@WebServlet(name = "DeleteCustomerController", value = "/admin/customer/delete")
+public class DeleteCustomerController extends HttpServlet {
+    private ICustomerService customerService;
 
     @Override
     public void init() throws ServletException {
         // Initialize the service implementation
-        employeeService = new EmployeeServiceImpl();
+         customerService = new CustomerServiceImpl();
     }
+
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         if (request.getSession().getAttribute("admin") == null) {
@@ -26,20 +27,25 @@ public class DeleteController extends HttpServlet {
             return;
         }
         try {
-            // Retrieve the employeeId from the request
+            // Retrieve the userId from the request
             int userId = Integer.parseInt(request.getParameter("userId"));
 
-            // Call the service to delete the employee
-            employeeService.delete(userId);
+            // Call the service to delete the customer
+            customerService.delete(userId);
 
-            // Redirect to the employee management page upon success
-            response.sendRedirect(request.getContextPath() + "/admin/employee");
+            // Redirect to the customer management page upon success
+            response.sendRedirect(request.getContextPath() + "/admin/customer");
 
         } catch (Exception e) {
             e.printStackTrace();
             // Forward the error details to an error page
-            request.setAttribute("errorMessage", "Failed to delete the employee. Please try again.");
+            request.setAttribute("errorMessage", "Failed to delete the customer. Please try again.");
             request.getRequestDispatcher("/errorPage.jsp").forward(request, response);
         }
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
     }
 }
