+++ src/main/java/cnpm/ergo/controller/Admin/Customer/UpdateCustomerController.java
@@ -1,5 +1,5 @@
 package cnpm.ergo.controller.Admin.Customer;
-import cnpm.ergo.entity.Customer;
+
 import cnpm.ergo.service.implement.CustomerServiceImpl;
 import cnpm.ergo.service.interfaces.ICustomerService;
 import jakarta.servlet.ServletException;
@@ -7,18 +7,20 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
 
-@WebServlet(name = "AddController", value = "/admin/customer/add")
-public class AddController extends HttpServlet {
+import java.io.IOException;
+import cnpm.ergo.entity.Customer;
+
+@WebServlet(name = "UpdateEmployeeController", value = "/admin/customer/update")
+public class UpdateCustomerController extends HttpServlet {
     private ICustomerService customerService;
-    @Override
+
     public void init() throws ServletException {
-        // Initialize the service implementation
         customerService = new CustomerServiceImpl();
-}
+    }
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
     }
 
     @Override
@@ -29,31 +31,36 @@ public class AddController extends HttpServlet {
         }
         try {
             // Retrieve form data from the request
+            int userId = Integer.parseInt(request.getParameter("userId"));
             String name = request.getParameter("name");
             String email = request.getParameter("email");
             String phone = request.getParameter("phone");
             String address = request.getParameter("address");
             String gender = request.getParameter("gender");
+            String status = request.getParameter("status");
             String password = request.getParameter("password");
 
             // Validate input fields (optional, add your validation logic here)
 
-            // Create and populate the Customer object
-            Customer customer = new Customer();
-            customer.setName(name);
-            customer.setEmail(email);
-            customer.setPhone(phone);
-            customer.setAddress(address);
-            customer.setGender(gender);
-            customer.setPassword(password);
-            customer.setStatus("Active"); // Default status
-            customer.setIsDelete(false); // Mark as not deleted
+            // Retrieve the existing Customer object
+            Customer customer = customerService.getCustomerById(userId);
+            if (customer != null) {
+                // Update the Customer object with new values
+                customer.setName(name);
+                customer.setEmail(email);
+                customer.setPhone(phone);
+                customer.setAddress(address);
+                customer.setGender(gender);
+                customer.setStatus(status);
+                customer.setPassword(password); // Update password
 
-            customerService.insert(customer);
-            response.sendRedirect(request.getContextPath() + "/admin/customer");
-        } catch (Exception e) {
+                // Call the service to update the customer
+                customerService.update(customer);
+                response.sendRedirect(request.getContextPath() + "/admin/customer");
+            }
+        } catch (Exception e){
             e.printStackTrace();
-            request.setAttribute("errorMessage", "Failed to add the customer. Please try again.");
+            request.setAttribute("errorMessage", "Failed to update the customer. Please try again.");
             request.getRequestDispatcher("/errorPage.jsp").forward(request, response);
         }
     }
